requirements.txt
anyjson==0.3.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0
django-crispy-forms==1.8.1
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
Pillow==7.0.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytemperature==1.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
sqlparse==0.3.0
style==1.1.0
update==0.0.1
urllib3==1.25.7
wrapt==1.11.2
